export a more reasonable debugging api
 - import devkit.debugging;
 - remove _DEBUG, attach to devkit.debugging instead
